move some classes into own file for better overview
