+++ AsyJob.Lib/Jobs/ProgressStatus.cs
@@ -1,0 +1,7 @@
+﻿namespace AsyJob.Lib.Jobs
+{
+    public enum ProgressStatus
+    {
+        Created, Waiting, Running, Done, Error
+    }
+}
