+++ AsyJob.Lib/Runner/IJobRunner.cs
@@ -21,24 +21,4 @@ namespace AsyJob.Lib.Runner
         /// </summary>
         Task<Job?> GetJob(string jobId);
     }
-
-    public class JobRunnerService(IJobPool pool) : IJobRunner
-    {
-        private readonly IJobPool _pool = pool;
-
-        public void RunJob(Job job)
-        {
-            _pool.RunJob(job);
-        }
-
-        public Task<IEnumerable<Job>> GetJobs()
-        {
-            return _pool.FetchAll<Job>();
-        }
-
-        public Task<Job?> GetJob(string jobId)
-        {
-            return _pool.FetchJob<Job>(jobId);
-        }
-    }
 }
