+++ AsyJob/Jobs/JobRequestDto.cs
@@ -1,5 +1,4 @@
-﻿using AsyJob.Lib.Jobs;
-using System.Dynamic;
+﻿using System.Dynamic;
 
 namespace AsyJob.Jobs
 {
@@ -9,10 +8,5 @@ namespace AsyJob.Jobs
         public string Description { get; private set; } = description;
         public string JobType { get; private set; } = jobtype;
         public ExpandoObject? Input { get; private set; } = input;
-    }
-
-    public class JobResponseDto(Job job)
-    {
-        public Job Job { get; private set; } = job;
     }
 }
