+++ AsyJob/Jobs/JobResponseDto.cs
@@ -1,0 +1,10 @@
+﻿using AsyJob.Lib.Jobs;
+
+namespace AsyJob.Jobs
+{
+
+    public class JobResponseDto(Job job)
+    {
+        public Job Job { get; private set; } = job;
+    }
+}
